Working add, get, and projection

+++ packages/adapters/source_control/persistence/postgres/migrations/0001_initial.sql
@@ -1,11 +1,11 @@
 CREATE TABLE "Organization" (
-    id integer primary key,
+    id bigint primary key,
     name varchar
 );
 
 CREATE TABLE "PlatformAccount" (
-    id integer primary key,
-    organization_id integer references "Organization",
+    id bigint primary key,
+    organization_id bigint references "Organization",
     name varchar,
     platform_name varchar
 );
